experiments/v3_trifoodnet_research_snapshot/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trifoodnet-research-snapshot"
version = "0.1.0"
description = "Research snapshot for TriFoodNet with official PictSure integration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "torch>=2.1",
  "torchvision>=0.19",
  "transformers>=4.44",
]

[project.optional-dependencies]
official-pictsure = [
  "PictSure",
  "huggingface_hub>=0.24",
]
research = [
  "bitsandbytes>=0.49",
  "peft>=0.18",
  "psutil>=5.9",
  "wandb>=0.18",
]
dev = [
  "pytest>=8.0",
]

[project.scripts]
trifoodnet-pictsure = "run_official_pictsure:main"
trifoodnet-benchmark = "benchmark_runtime:main"
trifoodnet-report = "experiment_report:main"

[tool.setuptools]
py-modules = [
  "benchmark_runtime",
  "config_loader",
  "dataset_integration",
  "experiment_report",
  "experiment_logging",
  "item_processing",
  "losses",
  "metrics",
  "pipeline",
  "pictsure_baseline",
  "pictsure_official",
  "run_official_pictsure",
  "stage1_qwen",
  "stage2_sam",
  "stage3_icl",
  "train_joint",
]

[tool.setuptools.data-files]
"." = ["master_config.yaml"]
